+++ graphbrain/commands/create.py
@@ -1,7 +1,0 @@
-from graphbrain import hgraph
-
-
-def run(args):
-    print('creating hypergraph...')
-    hgraph(args.hg)
-    print('done.')